Handling Oddly-Formed plists Better
Handle plists where the Office 2008 recent file list is not a perfect
sequential order from one to whatever.  Office likely wouldn’t create a
non-sequential set of recently used files, but if the list was edited
by hand or script it would have failed to be cleaned with previous
versions of this script.